to_do/requirements.txt
asgiref==3.4.1
asttokens==2.0.5
beautifulsoup4==4.10.0
black==21.11b1
bootstrap4==0.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==21.1
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-database-constraints==2.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==2
drf-yasg==1.20.0
drf-yasg2==1.19.4
executing==0.8.2
icecream==2.1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
loguru==0.5.3
Markdown==3.3.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.2
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycparser==2.21
Pygments==2.10.0
pyparsing==2.4.7
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
